Enable ls() to search for multiple distributions.
This will require a hashref for a return value.  So let's make things more
internally consistent and have ls() return a reference whether we're examining
single or multiple distributions.  In the single case, return a one-element
array ref.  In the multiple case, return a hash ref keyed on distribution name
with the element being an array ref of versions found on the server.